requirements.txt
python-telegram-bot==21.*
google-genai
sqlalchemy
python-dotenv
